battleLogic
Got the initiative advancing for each unit while making the units act in-editor

+++ Assets/Scripts/BaseUnit.cs
@@ -6,11 +6,12 @@ public class BaseUnit : MonoBehaviour
 {
     public string unitName;
 
+    // We should at least set speed to be non-negative. The BattleStateMachine's while loops will go infinite if initiative doesn't advance to the threshold.
     public float currentHP, maxHP, baseATK, currentATK, baseDEF, currentDEF, speed, stateCharge;
     public int fireTokens, waterTokens, earthTokens, skyTokens;
     public bool dualState;
 
-    public float initiative, simulatedInitiative;
+    public double initiative, simulatedInitiative;
 
     public List<BaseAttack> attackList = new List<BaseAttack>();
 }
